backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.4
pydantic-settings==2.9.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
email-validator==2.2.0
scikit-learn==1.7.0
pandas==2.3.0
numpy==2.2.6
joblib==1.5.1
reportlab==4.4.1
